backend/pyproject.toml
[tool.poetry]
name = "sora-backend"
version = "0.1.0"
description = "FastAPI backend for Sora video generator (migrated from Java Spring Boot)"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
httpx = "^0.27.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
tenacity = "^9.0.0"
azure-identity = "^1.17.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.8"
ruff = "^0.5.0"
pytest-httpx = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
